requirements.txt
backports.functools-lru-cache==1.5
cloudpickle==0.6.1
cycler==0.10.0
dask==0.20.0
decorator==4.3.0
easydict==1.9
funcsigs==1.0.2
kiwisolver==1.0.1
matplotlib==2.2.3
mock==2.0.0
networkx==2.2
nltk==3.3
numpy==1.15.4
pbr==5.1.0
Pillow==5.3.0
pkg-resources==0.0.0
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
subprocess32==3.5.3
tensorboard==1.0.0a4
toolz==0.9.0
torch==0.4.1
torchfile==0.1.0
torchvision==0.2.1
Werkzeug==0.14.1
